Separated PUID into it's own new file.
