+++ assets/models/coin.mtl
@@ -1,0 +1,12 @@
+# Blender 4.0.1 MTL File: 'None'
+# www.blender.org
+
+newmtl Coin
+Ns 0.000000
+Ka 0.673575 0.673575 0.673575
+Kd 0.800513 0.401362 0.017388
+Ks 0.500000 0.500000 0.500000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 3
