+++ core/include/PurrfectEngine/PurrfectEngine.hpp
@@ -8,13 +8,19 @@
 
 #include <string.h>
 #include <vector>
+#include <algorithm>
+#include <optional>
+#include <unordered_map>
 
 #include "PurrfectEngine/extensions.hpp"
 
 #include "PurrfectEngine/transform.hpp"
 #include "PurrfectEngine/camera.hpp"
-#include "PurrfectEngine/scene.hpp"
 #include "PurrfectEngine/audio.hpp"
+
+#include "PurrfectEngine/PUID.hpp"
+#include "PurrfectEngine/object.hpp"
+#include "PurrfectEngine/scene.hpp"
 
 namespace PurrfectEngine {
