+++ core/src/object.cpp
@@ -6,14 +6,6 @@
 
 namespace PurrfectEngine {
 
-  static std::random_device sRandomDevice;
-  static std::mt19937_64 sEngine(sRandomDevice());
-  static std::uniform_int_distribution<uint32_t> sUniformDistribution;
-
-  PUID::PUID():
-    mId(sUniformDistribution(sEngine))
-  {}
-
   purrComponent::purrComponent()
   {}
